\name{alberta}
\alias{alberta}
\docType{data}
\title{Alberta Survey}
\description{
Data from the 2013 Alberta Survey administered by the Population Research Lab at the University of Alberta. For complete codebook, see here \url{https://dataverse.library.ualberta.ca/dataset.xhtml?persistentId=doi:10.7939/DVN/10573}.}
\usage{data(alberta)}
\format{
  A data frame with 2751 observations on the following 56 variables.  Each of the policy statements is coded 1 (strongly disagree), 2 (disagree), 3 (agree), 4 (strongly agree), 8 (don't know)
  \describe{
\item{respnum}{Respondent number}
\item{primary}{primary}
\item{strata}{Area of the province}
\item{sex}{Gender of respondent}
\item{d1}{Your current debt/financial situation}
\item{d2}{Past two years, completely missed a payment/been at least two months late in pay}
\item{d3}{Past two years, consulted a debt consolidation or credit-counseling firm}
\item{d4}{Past two years, been denied for a loan or credit card due to your credit rating}
\item{d5}{Ever declared (filed for) bankruptcy, or a bankruptcy alternative jointly or alo}
\item{d6}{Currently have a checking or savings account at a bank or credit union}
\item{d7}{Currently hold or make payments on any student loans in your name}
\item{d8}{Have any credit cards in your name}
\item{d9}{Number of credit cards you have}
\item{d10}{How often do you pay the entire or full monthly balance on all of your credit ca}
\item{e1}{While growing up, before your 18th birthday: frequently experienced verbal insul}
\item{e2}{While growing up, before your 18th birthday: ever injured or bruised from physic}
\item{e3}{While growing up, before your 18th birthday: experienced inappropriate sexual ad}
\item{e4}{While growing up, before your 18th birthday: ever witnessed your mother or stepm}
\item{e5}{While growing up, before your 18th birthday: part of a household where someone a}
\item{e6}{While growing up, before your 18th birthday: part of a household where someone w}
\item{e7a}{While growing up, before your 18th birthday: part of a household where someone w}
\item{e8}{While growing up, before your 18th birthday: your parents separated or divorced}
\item{e9}{While growing up, before your 18th birthday: rate your level of stress of your c}
\item{e10}{Rate your emotional health today}
\item{e11}{Rate your physical health today}
\item{e12}{I have family and friends who help me feel safe, secure and happy}
\item{e13}{I usually expect things to go my way}
\item{e14_1}{Diagnosed with high blood pressure as an adult}
\item{e14_2}{Diagnosed with diabetes as an adult}
\item{e14_3}{Diagnosed with  irritable bowel syndrome/Crohn's disease as an adult}
\item{e14_4}{Diagnosed with chronic pain as an adult}
\item{e14_5}{Diagnosed with backache as an adult}
\item{e14_6}{Diagnosed with asthma as an adult}
\item{e14_7}{Diagnosed with allergies as an adult}
\item{e14_8}{Diagnosed with chronic fatigue syndrome/ fibromyalgia as an adult}
\item{e14_9}{Diagnosed with anxiety disorder as an adult}
\item{e14_10}{Diagnosed with MDD (major depressive disorder) as an adult}
\item{e14_11}{Diagnosed with alcohol dependency problems as an adult}
\item{e14_12}{Diagnosed with drug dependency problems as an adult}
\item{e14_13}{Diagnosed with other health problems as an adult}
\item{e14_15}{Diagnosed with no health problems as an adult}
\item{e14_16}{Treatment for high blood pressure}
\item{e14_17}{Treatment for diabetes}
\item{e14_18}{Treatment for irritable bowel syndrome/Crohn's disease}
\item{e14_19}{Treatment for chronic pain}
\item{e14_20}{Treatment for backache}
\item{e14_21}{Treatment for asthma}
\item{e14_22}{Treatment for allergies}
\item{e14_23}{Treatment for chronic fatique syndrome/fibromyalgia}
\item{e14_24}{Treatment for anxiety disorder}
\item{e14_25}{Treatment for MDD (major depressive dsorder)}
\item{e14_26}{Treatment for alcohol dependency problems}
\item{e14_27}{Treatment for drug dependency problems}
\item{e14_28}{Treatment for other specified health problems}
\item{ft1}{How much do you know about the Temporary Foreign Worker Program?}
\item{ft2}{Skill and labour shortages in Canada should be met by...}
\item{ft3}{Employers in Alberta should be able to hire Temporary Foreign Workers for...}
\item{ft4}{Temporary foreign workers are needed to fill jobs in the Alberta labour market}
\item{ft5}{Temporary Foreign Workers are taking jobs away from Albertans}
\item{ft6}{Most Temporary Foreign Workers are members of visible minority groups (non-white}
\item{ft7}{It is OK for Alberta Workplace laws and standards to be less strict for Temporar}
\item{ft8}{Temporary Foreign Workers should be given the exact same workplace rights as Can}
\item{ft9}{It is OK if Temporary Foreign Workers are paid less than Canadian workers perfor}
\item{ft10}{All Temporary Foreign Workers should have the opportunity to obtain permanent re}
\item{g1}{Leading cause of death for Albertans under the age of 45}
\item{g2}{Thinking about the different health problems facing Albertans today, how serious}
\item{g3}{Injuries are...}
\item{g4}{Chances that you will visit an emergency room because of an injury in the next y}
\item{g5}{Chances that you will visit an emergency room because of a motor vehicle collisi}
\item{g6a}{How preventable are poisonings?}
\item{g6b}{How preventable are drownings?}
\item{g6c}{How preventable are falls from ladders?}
\item{g6d}{How preventable are workplace injuries?}
\item{g6e}{How preventable are head injuries from not wearing a helmet while bicycle riding}
\item{g6f}{How preventable are motor vehicle crashes from driving while distracted (e.g., c}
\item{g6g}{How preventable are motor vehicle crashes from falling asleep at the wheel?}
\item{g6h}{How preventable are injuries resulting from jaywalking?}
\item{g6i}{How preventable are burns in the home from fire, flames & hot substances?}
\item{g6j}{How preventable are operating machinery or a motor vehicle while under the influ}
\item{g6k}{How preventable are suicides?}
\item{g7}{Thinking about the different injury issues, for example, motor vehicle collision}
\item{g8}{Thinking about the different injury issues, for example, motor vehicle collision}
\item{h1}{Do you believe that alcohol use contributes to health problems in Alberta?}
\item{h2}{Do you believe alcohol use contributes to injuries in Alberta?}
\item{h3_1}{Individuals have a responsibility to reduce alcohol-related injuries}
\item{h3_2}{Families have a responsibility to reduce alcohol-related injuries}
\item{h3_3}{Communities have a responsibility to reduce alcohol-related injuries}
\item{h3_4}{Government has a responsibility to reduce alcohol-related injuries}
\item{h3_5}{Health care has a responsibility to reduce alcohol-related injuries}
\item{h3_6}{Police have a responsibility to reduce alcohol-related injuries}
\item{h3_7}{Workplaces have a responsibility to reduce alcohol-related injuries}
\item{h3_8}{Alcohol industry/producers have a responsibility to reduce alcohol-related injur}
\item{h3_9}{Hospitality industry has a responsibility to reduce alcohol-related injuries}
\item{h4}{Best way for the government to reduce alcohol problems}
\item{h5}{Had at least one drink of any alcoholic beverage during the past 30 days?}
\item{h6}{Number of DAYS had at least one drink of any alcoholic beverage during the past }
\item{h7}{On the days when you drank, number of DRINKS on average during the past 30 days?}
\item{h8}{Considering all types of alcoholic beverages, number of times during the past 30}
\item{k1a}{Your present employment status}
\item{k3a}{Number of adults (including yourself) living in household}
\item{k3b}{Number of children living in household}
\item{k3c}{Total number of persons in household}
\item{age}{Your age}
\item{agex}{Recoded age in six groups}
\item{k5a}{Current marital status}
\item{k6}{Highest level of education}
\item{k6group}{Recoded education into three groups}
\item{k7}{Total years of schooling}
\item{k8a}{Your religion}
\item{mrelig}{Recoded religion into four groups}
\item{canb}{Were you born in Canada?}
\item{abb}{Were you born in Alberta?}
\item{cgrp}{Are you...?}
\item{k10}{Would you say that you (and your family) are BETTER OFF, just the SAME, or WORSE}
\item{k11}{A year from now,  do think you you (and your family) will be BETTER OFF, just ab}
\item{k12a}{Total  Household income for the past year before taxes and deductions}
\item{k13}{Do you (or your spouse/partner/parents) presently own or rent your residence?}
\item{k16a}{If an election was held today, how would you vote federally?}
\item{k16b}{If an election was held today, how would you vote provincially?}
\item{k17}{Postal code}
\item{wtx}{Weights from Cansim estimate for 2012}
\item{wtx2}{Weights by Age (3) and Gender by Area Cansim Estimate}
\item{agegroupxx}{Three agegroups used in weighting variable wtx2}
\item{b1}{China's increasing economic strength benefits Alberta}
\item{b2}{Alberta should build stronger ties with China}
\item{b3}{China is important as an export market for Alberta goods and services}
\item{b4}{Alberta should decrease its economic reliance on the US market}
\item{b5}{Alberta should diversify its economy by trading more with Asia}
\item{b6}{Alberta should support the building of better infrastructure to transport energy}
\item{b7}{Alberta should welcome Chinese investment in the province}
\item{b9}{Alberta should welcome Chinese investment in energy and other resource sectors}
\item{b10}{China will play an increasingly significant role in the future opportunities of }
\item{b11}{Increased tourism from China is good for Alberta}
\item{b12}{The ability to speak Chinese will become more important to Albertans}         }
}
\source{
Population Research Laboratory, 2015, "Alberta Survey, 2013", https://doi.org/10.7939/DVN/10573, UAL Dataverse, V3
}
\keyword{datasets}
