\name{histDiscrete}
\alias{histDiscrete}
\title{Create a histogram from a discrete variable}
\description{A histogram for a discrete variable is basically a bar chart with no space between the bars.}
\usage{
histDiscrete(x, ...)
}
\arguments{
  \item{x}{A vector of integer values or a \code{factor}.}
  \item{\dots}{Other arguments to be passed down to \code{\link{hist}}.}
}
\value{
A histogram produced with \code{\link{hist}}
}
