\name{sumStats}
\alias{sumStats}
\title{Summary Statistics}
\description{
Producessummary statistics for a set of variables.}
\usage{
summaryStats(data, vars)
}
\arguments{
  \item{data}{A \code{data.frame} or \code{tibble}.}
  \item{vars}{A vector of character strings giving the names of variables to be summarized}
}
\value{
A matrix giving summary statistics
}
