\name{uwo4419-package}
\alias{uwo4419-package}
\alias{uwo4419}
\docType{package}
\title{
\packageTitle{uwo4419}
}
\description{
\packageDescription{uwo4419}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{uwo4419}
\packageIndices{uwo4419}
}
\author{
\packageAuthor{uwo4419}

Maintainer: \packageMaintainer{uwo4419}
}
